+++ CHANGES.md
@@ -16,6 +16,7 @@
 
 - libwebrtc の internal tracer 利用有無を `PeerConnectionOption` で指定可能にした
 - シグナリング connect メッセージに `sdk_type`, `sdk_version` と `user_agent` を追加した
+- シグナリング connect メッセージに `audio.opus_params` を追加した
 
 ## develop